chore(misc): shuffle some code
